requirements.txt
opencv-python
huggingface_hub
python-dotenv